# specs/001-nutrition-recipe-planner/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Cookfully API
  version: 0.2.0
  description: >-
    Canonical HTTP contract for recipes, honest nutrition estimates, goals, meal plans,
    grocery lists, exports, agent access, and later goal-aware suggestions. Public decimal
    values are canonical strings. Estimated nutrition and suggestion projections are planning
    aids, not medical advice. MCP tools compose the same application commands and queries.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /pantry-items:
    get:
      x-release: expansion-p6
      tags: [Pantry]
      operationId: listPantryItems
      responses:
        "200":
          description: Pantry inventory with exact quantities and reviewable matches.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PantryItem" }
    post:
      x-release: expansion-p6
      tags: [Pantry]
      operationId: createPantryItem
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PantryItemWrite" }
      responses:
        "201":
          description: Pantry item created; ambiguous matches remain reviewable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PantryItem" }
        "422":
          $ref: "#/components/responses/ValidationError"
  /pantry-items/{itemId}:
    patch:
      x-release: expansion-p6
      tags: [Pantry]
      operationId: updatePantryItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IfMatchVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PantryItemWrite" }
      responses:
        "200":
          description: Quantity or manual food-reference match updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PantryItem" }
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      x-release: expansion-p6
      tags: [Pantry]
      operationId: deletePantryItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IfMatchVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Pantry item removed when no applied deduction depends on it.
        "409":
          $ref: "#/components/responses/Conflict"
  /pantry/recipe-matches:
    get:
      x-release: expansion-p6
      tags: [Pantry]
      operationId: findMakeableRecipes
      parameters:
        - name: query
          in: query
          schema: { type: [string, "null"], maxLength: 200 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: Fully makeable recipes first, then partial matches with missing ingredients.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PantryRecipeMatch" }
  /meal-plans/{weekStart}/grocery-list/pantry-deductions:
    post:
      x-release: expansion-p6
      tags: [Pantry, Grocery]
      operationId: applyPantryDeductions
      parameters:
        - $ref: "#/components/parameters/WeekStart"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PantryDeductionApply" }
      responses:
        "200":
          description: Only exact safe matches deducted; every result remains visible and reversible.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PantryDeduction" }
        "409":
          $ref: "#/components/responses/Conflict"
  /pantry-deductions/{deductionId}:
    delete:
      x-release: expansion-p6
      tags: [Pantry, Grocery]
      operationId: reversePantryDeduction
      parameters:
        - name: deductionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/IfMatchVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Deducted pantry and grocery quantities restored exactly.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PantryDeduction" }
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cookfully_session
    bearerToken:
      type: http
      scheme: bearer
  parameters:
...
    WeekStart:
      name: weekStart
      in: path
      required: true
      schema: { type: string, format: date }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
...
  responses:
...
    Conflict:
      description: Stale version, duplicate idempotency key with different input, or invalid transition.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: One or more fields are invalid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
    Decimal6:
      type: string
      description: Canonical non-negative fixed decimal with at most six fractional places and no exponent.
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]{1,6})?$'
...
    RatioDecimal:
      type: string
      description: Canonical fixed decimal from zero through one.
      pattern: '^(0(\.[0-9]{1,6})?|1(\.0{1,6})?)$'
...
    PantryItemWrite:
      type: object
      additionalProperties: false
      required: [displayName, quantity, unit]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 240 }
        quantity: { $ref: "#/components/schemas/Decimal6" }
        unit: { type: string, minLength: 1, maxLength: 80 }
        expiresOn: { type: [string, "null"], format: date }
        foodReferenceId: { type: [string, "null"], format: uuid }
        ownerFoodId: { type: [string, "null"], format: uuid }
    PantryItem:
      allOf:
        - $ref: "#/components/schemas/PantryItemWrite"
        - type: object
          required: [id, normalizedFoodName, matchStatus, matchConfidence, version]
          properties:
            id: { type: string, format: uuid }
            normalizedFoodName: { type: string }
            matchStatus: { type: string, enum: [unmatched, proposed, matched, manual] }
            matchConfidence: { oneOf: [{ $ref: "#/components/schemas/RatioDecimal" }, { type: "null" }] }
            version: { type: integer, minimum: 1 }
    PantryRecipeMatch:
      type: object
      additionalProperties: false
      required: [recipeId, recipeTitle, availability, coverageRatio, missingIngredients]
      properties:
        recipeId: { type: string, format: uuid }
        recipeTitle: { type: string }
        availability: { type: string, enum: [full, partial, none] }
        coverageRatio: { $ref: "#/components/schemas/RatioDecimal" }
        missingIngredients: { type: array, items: { type: string } }
    PantryDeductionApply:
      type: object
      additionalProperties: false
      required: [expectedGroceryListVersion]
      properties:
        expectedGroceryListVersion: { type: integer, minimum: 1 }
        groceryItemIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
    PantryDeduction:
      type: object
      additionalProperties: false
      required: [id, pantryItemId, groceryItemId, pantryQuantity, pantryUnit, groceryQuantity, groceryUnit, assumption, status, version]
      properties:
        id: { type: string, format: uuid }
        pantryItemId: { type: string, format: uuid }
        groceryItemId: { type: string, format: uuid }
        pantryQuantity: { $ref: "#/components/schemas/Decimal6" }
        pantryUnit: { type: string }
        groceryQuantity: { $ref: "#/components/schemas/Decimal6" }
        groceryUnit: { type: string }
        assumption: { type: string }
        status: { type: string, enum: [applied, reversed] }
        appliedAt: { type: string, format: date-time }
        reversedAt: { type: [string, "null"], format: date-time }
        version: { type: integer, minimum: 1 }
...